zeta-agent-demo/backend/requirements.txt
fastapi
uvicorn
qwen-agent
web3
python-dotenv
pydantic
